* Shell-scripts -> update [ build.sh && run.sh ]
The “top layer”, “logic” and output of the audio stream to the speakers in the engine will be written in Rust. This is because this language is more secure than C. The core of the engine will be written in C, because it is easier to implement various functions for sound processing and low-level optimizations in this language.